The all-clusters app relies on the af-gen-event.h file for the EMBER_AF_GENERATED_EVENTS define, along with its context. However, the file is not automatically generated by zap. Thus other sample apps that rely on level control don't perform the transition.
This define enables the cluster events/ticks that cause a transition from one state to another on time-dependent clusters such as level control.

This change to af-event.cpp/h deprecates the define and expects the cluster will register itself during its initialization. This self-init was implemented on the level control cluster.

Change-Id: I3b4e56beacad4741d8462d97af6bb694c810d4a9